feat: config changes
