+++ indica/configs/operators/__init__.py
@@ -1,0 +1,2 @@
+"""Operator-specific configuration modules."""
+
